backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
et_xmlfile==2.0.0
firebase-admin==6.4.0
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==2.14.0
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
gunicorn==21.2.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==1.26.4
openpyxl==3.1.2
packaging==26.0
pandas==2.1.4
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyparsing==3.3.2
PyPDF2==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2026.1.post1
requests==2.33.1
setuptools==82.0.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.8
wheel==0.46.3
